ui expert: knowledge about forms architecture
